Moved major functions to separate class
